feat: refactor lead API to enhance resume handling and email verification
