+++ workers/lead-api/constants.ts
@@ -1,0 +1,3 @@
+export const EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+export const RESUME_TOKEN_TTL_SECONDS = 7 * 24 * 60 * 60;
+export const DEFAULT_RESUME_RESEND_COOLDOWN_MINUTES = 15;
